only fill faq list if the table exists

+++ config/initializers/faq.rb
@@ -1,9 +1,11 @@
 # This initializer sets up the faq table
 
-# First destroy all records
-Faq.delete_all
+if Faq.table_exists?
+  # First destroy all records
+  Faq.delete_all
 
-# Now create the FAQs from JSON
-JSON.parse(File.read("#{Rails.root}/config/faq.json")).each do |key, value|
-  Faq.create(:question => value['q'], :answer => value['a'])
+  # Now create the FAQs from JSON
+  JSON.parse(File.read("#{Rails.root}/config/faq.json")).each do |key, value|
+    Faq.create(:question => value['q'], :answer => value['a'])
+  end
 end
